+++ lib/apm_v5/plugins/plugin_registry.ex
@@ -79,7 +79,8 @@ defmodule ApmV5.Plugins.PluginRegistry do
     ApmV5.Plugins.Ports.PortsPlugin,
     ApmV5.Plugins.Usage.UsagePlugin,
     ApmV5.Plugins.Devops.DevopsPlugin,
-    ApmV5.Plugins.Alerting.AlertingPlugin
+    ApmV5.Plugins.Alerting.AlertingPlugin,
+    ApmV5.Plugins.SimpleAgents.SimpleAgentsPlugin
   ]
 
   @impl true